Document-processing-function/requirements.txt
google-cloud-documentai>=2.29.0,<3.0.0
google-cloud-firestore>=2.15.0,<3.0.0
google-cloud-pubsub>=2.20.0,<3.0.0
